Move Deferred collection free functions to protocol extensions
