+++ Deferred/Deferred.swift
@@ -264,49 +264,3 @@ extension Deferred {
         return flatMap { t in other.map { u in (t, u) } }
     }
 }
-
-/**
-Compose a number of deferred values into a single deferred array.
-
-:param: deferreds Any collection whose elements are themselves deferred values.
-:return: A deferred array that is determined once all the given values are
-determined, in the same order.
-**/
-public func all<Value, Collection: CollectionType where Collection.Generator.Element == Deferred<Value>>(deferreds: Collection) -> Deferred<[Value]> {
-    let array = Array(deferreds)
-    if array.isEmpty {
-        return Deferred(value: [])
-    }
-
-    let combined = Deferred<[Value]>()
-    let group = dispatch_group_create()
-
-    for deferred in array {
-        dispatch_group_enter(group)
-        deferred.upon { _ in
-            dispatch_group_leave(group)
-        }
-    }
-
-    dispatch_group_notify(group, genericQueue) {
-        let results = array.map { $0.value }
-        combined.fill(results)
-    }
-
-    return combined
-}
-
-/**
-Choose the deferred value that is determined first.
-
-:param: deferreds Any collection whose elements are themselves deferred values.
-:return: A deferred value that is determined with the first of the given
-deferred values to be determined.
-**/
-public func any<Value, Sequence: SequenceType where Sequence.Generator.Element == Deferred<Value>>(deferreds: Sequence) -> Deferred<Value> {
-    let combined = Deferred<Value>()
-    for d in deferreds {
-        d.upon { t in combined.fill(t, assertIfFilled: false) }
-    }
-    return combined
-}
